고치느라 시간 다 보냇네
