+++ frontend/src/App.jsx
@@ -13,15 +13,14 @@ import { MemberLogin } from "./feature/member/MemberLogin.jsx";
 import { AuthenticationContextProvider } from "./common/AuthenticationContextProvider.jsx";
 import { MapDetail } from "./feature/map/MapDetail.jsx";
 import FullFilterKakaoMap from "./feature/map/FullFilterKakaoMap";
-import { ReviewAdd } from "./feature/map/ReviewAdd.jsx";
+import { ReviewAdd } from "./feature/review/ReviewAdd.jsx";
 import { Chatbot } from "./feature/openai/Chatbot";
-import { ReviewEdit } from "./feature/map/ReviewEdit.jsx";
-import { ReviewListMini } from "./feature/board/ReviewListMini.jsx";
+import { ReviewEdit } from "./feature/review/ReviewEdit.jsx";
+import { LatestReviewsList } from "./feature/review/LatestReviewsList.jsx";
 import ServicePage from "./feature/service/ServicePage.jsx";
 import ServiceListPage from "./feature/service/ServiceListPage.jsx";
 import { MyReview } from "./feature/review/MyReview.jsx";
 // import { MemberLogout } from "./feature/member/MemberLogout.jsx";
-
 import { KakaoCallback } from "./feature/member/KakaoCallback.jsx";
 
 function App() {
@@ -52,7 +51,7 @@ function App() {
 
             <Route path="/review/my" element={<MyReview />} />
             <Route path="/review/edit/:id" element={<ReviewEdit />} />
-            <Route path="/review/latest" element={<ReviewListMini />} />
+            <Route path="/review/latest" element={<LatestReviewsList />} />
 
             <Route path="/chatbot" element={<Chatbot />} />
