+++ frontend/src/feature/main/BoardListMini.jsx
